Move launching task to a helper function so it can be more easily shared.
